Add a renderer module to move the rendering functions into it
