+++ src/scenes.rs
@@ -1,9 +1,10 @@
-use raytracing::materials::{Dielectric, LambertianMaterial, Metal};
-use raytracing::objects::{HittableList, Sphere};
-use raytracing::{
-    utils::{random, random_in_unit_interval, PI},
-    Camera, Color, Point, Vec3,
+use crate::materials::{Dielectric, LambertianMaterial, Metal};
+use crate::objects::{HittableList, Sphere};
+use crate::{
+    camera::Camera
 };
+use crate::utils::{random, random_in_unit_interval, PI};
+use crate::vec3::{Vec3, Color, Point};
 use std::rc::Rc;
 
 pub(crate) fn scene_with_dielectric_and_shiny_sphere() -> HittableList {
